Incorporate _solve_inequality into solve.
_solve_inequality is a very simple solver for linear inequalities
in several variables.

Extract it from transforms.py and move it to solvers/inequalities.py.
Use it in reduce_inequalities.
Add tests.
Also import _solve_inequality in transforms.py again.
